+++ lua/plugins/config/lsp/init.lua
@@ -48,23 +48,6 @@ return function(use)
 		end,
 	})
 	use({
-		'vuki656/package-info.nvim',
-		requires = 'MunifTanjim/nui.nvim',
-		ft = { 'json' },
-		config = function()
-			local present, package_info = pcall(require, 'package-info')
-			if not present then
-				return
-			end
-			package_info.setup({
-				colors = {
-					up_to_date = '#ffd042',
-					outdated = '#db0d0d',
-				},
-			})
-		end,
-	})
-	use({
 		'weilbith/nvim-code-action-menu',
 		cmd = 'CodeActionMenu',
 	})